Move `listener-collection.js` to `src/shared`
`src/shared/port-rpc.js` depended on
`src/annotation/util/listener-collection.js`. Files in the `src/shared`
folder should not have dependencies on `src/annotation`.

`ListenerCollection` is a very general utility and the placement in
`src/shared` seems correct.

I took the opportunity to reorder the imports according to our
conventions.
