+++ src/sidebar/util/observe-element-size.js
@@ -1,4 +1,4 @@
-import { ListenerCollection } from '../../annotator/util/listener-collection';
+import { ListenerCollection } from '../../shared/listener-collection';
 
 /**
  * Watch for changes in the size (`clientWidth` and `clientHeight`) of
